+++ CHANGES.rst
@@ -16,6 +16,12 @@ Changelog of flooding-worker
 
 - Added functionality to start workers as subprocess.
 
+- Added success field to WorkflowTask model.
+
+- Added function to define status of workflow.
+
+- Added unit tests.
+
 0.6.2 (2012-06-28)
 ------------------